Generalize test to newer Sphinx versions
Newer versions of Sphinx changed how javascript is included in HTML. In
older version, the script tag had a type="application/javascript"
attribute, in newer, versions, no type is declared. This is because in
HTML5, the type became optional:
https://www.w3schools.com/tags/att_script_type.asp

Some of the tests that were looking for whether the
doctr-versions-menu.js was properly included in the rendered Sphinx
output were written under the assumption that the "type" attribute would
be there. Now, we only check for any reference to
"_static/doctr-versions-menu.js", which should work with any version of
Sphinx/HTML.