Inherit `NonVersionFolderName` from `_BaseVersion`
This allows us to go back to using `packaging` instead of `packvers` as
the dependency for parsing folder names/version numbers.

We'll have to depend on the internal `packaing.version._BaseVersion`,
but we can test the behavior that we rely on in case `packaging` makes a
breaking change to that behavior.